Youtube_api/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-admin-numeric-filter==0.1.6
idna==2.8
importlib-metadata==1.5.0
inflect==4.1.0
isodate==0.6.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
yt-search==1.1.0.post1
zipp==2.2.0
